Restructuring stack vm project.
