+++ stackgvm/source/Makefile.x86_linux
@@ -1,0 +1,14 @@
+# Project: ExVM Multiplatform
+
+# Target
+BIN      = bin/gamevm_x86
+
+# Compiler settings
+CXXFLAGS = -O3 -Wall -W -m32 -march=native -mfpmath=sse -fomit-frame-pointer -fno-exceptions -fexpensive-optimizations -D_VM_HOST_OS=_VM_HOST_LINUX_INTEL_32
+
+# Makefile settings
+ARCH     = x86_linux
+MEXT     = $(ARCH)
+
+include interpreter.make
+
