+++ stackgvm/source/test_interpreter.cpp
@@ -1,0 +1,9 @@
+#include <cstdio>
+#include "include/gvm_core.hpp"
+
+using namespace GVM;
+
+int main() {
+  std::printf("Max Opcode %d\n", Opcode::_MAX);
+  return 0;
+}
